fix: :bug: The my page will now work without chrashing on some occations
The myPage works now as intended, two bugs. First: surunded the blocked user with a if, in cases the user loged inn status iss laging behind the site rendering. Second: The addUser methode was not updatet in all places where it was used, so the blocking and unblocking was not working

#95 @DanielNHansen @mefriis